DEBUG Non-AVX code
Fixed minor issues with the non-AVX code (wrong function headers,
etc...).

+++ src/preCondGradFloatC.c
@@ -10,7 +10,7 @@
 #include <assert.h>
 
 
-void preCondGradFloatC(int** elementList, int** elementListSize, int* numEntries, int* maxElement, int* idxEntries, float* X, float* grid, unsigned short int* YIdx, int *numPointsPerBox, float* boxEvalPoints, int numBoxes, float* a, float* b, float gamma, float weight, float* delta, int N, int M, int dim, int nH, int MBox)
+void preCondGradFloatC(int** elementList, int** elementListSize, int* numEntries, int* maxElement, int* idxEntries, float* X, float* grid, unsigned short int* YIdx, int *numPointsPerBox, float* boxEvalPoints, int numBoxes, double* a, double* b, float gamma, float weight, float* delta, int N, int M, int dim, int nH, int MBox)
 {
 	/* initialize elementList */
 	static const int elementListIncrement = 10000000;
